requirements.txt
langchain
langchain-cli
langchain-community
langchain_pinecone
pinecone-client
python-dotenv
pypdf
pytest
pandas
tqdm
scikit-learn